Remove the TypeKey concept from the base interfaces
The TypeKey concept was added for the COM support. Now that we have an alternative solution with generic attributes that we can get from a System.Type and we already need the type instance, let's just provide that and allow each unmanaged type system projection handle things how they see fit.