Provide CCW class support via the same pattern as our RCW support.
This PR is an alternative to #6.

This makes the support pluggable in the same manner as our interface support through a new method on the strategy for discovering interface details. It also generates the code in a manner that is better for trimming (the linker and ILC won't remove branches like `if (obj.GetType() == typeof(X))` if `X` is never instantiated as far as I know).

This design ends up using some additional reflection as CCW construction will require one more custom attribute lookup with this model than the alternative, but it also ensures that we don't need to generate checks for any interfaces the user implements that we know are not exposed to COM.